Backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# AWS Lambda Adapter (optional - only for Lambda deployment)
mangum==0.17.0

# Database
supabase==2.3.4

# AWS S3
boto3==1.34.34

# Google Gemini AI
google-generativeai==0.3.2

# Redis (for coordinator notes)
redis==5.0.1

# Resend (for email notifications)
resend==2.23.0

# HTTP Requests (for HuggingFace API)
requests==2.31.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0
